fix: clearly separate Actor related code and rest of the codebase. Only export ApifyMcpServer from the package
