+++ src/actor/README.md
@@ -1,0 +1,6 @@
+# Actor
+
+Code related to Apify Actor called Actors-MCP-Server.
+This Actor will be deprecated in favor of Apify MCP Server, therefore we are keeping it separate from the main codebase.
+
+The only exception is the `src/main.ts` file that also belongs to the Actor.
