+++ src/actor/server.ts
@@ -1,18 +1,21 @@
 /*
- * This file contains the express server implementation used for standby Actor mode.
+ * Express server implementation used for standby Actor mode.
  */
 
 import { SSEServerTransport } from '@modelcontextprotocol/sdk/server/sse.js';
 import type { Request, Response } from 'express';
 import express from 'express';
 
-import { HEADER_READINESS_PROBE, Routes } from './const.js';
-import { log } from './logger.js';
-import { type ApifyMcpServer } from './mcp-server.js';
+import log from '@apify/log';
 
-export function createExpressApp(host: string,
+import { HEADER_READINESS_PROBE, Routes } from './const.js';
+import { type ApifyMcpServer } from '../mcp-server.js';
+import { getActorRunData, processParamsGetTools } from './utils.js';
+
+export function createExpressApp(
+    host: string,
     mcpServer: ApifyMcpServer,
-    additionalData?: object): express.Express {
+): express.Express {
     const HELP_MESSAGE = `Connect to the server with GET request to ${host}/sse?token=YOUR-APIFY-TOKEN`
         + ` and then send POST requests to ${host}/message?token=YOUR-APIFY-TOKEN`;
 
@@ -29,11 +32,14 @@ export function createExpressApp(host: string,
             }
             try {
                 log.info(`Received GET message at: ${Routes.ROOT}`);
-                await mcpServer.processParamsAndUpdateTools(req.url);
+                const tools = await processParamsGetTools(req.url);
+                if (tools) {
+                    mcpServer.updateTools(tools);
+                }
                 res.setHeader('Content-Type', 'text/event-stream');
                 res.setHeader('Cache-Control', 'no-cache');
                 res.setHeader('Connection', 'keep-alive');
-                res.status(200).json({ message: `Actor is using Model Context Protocol. ${HELP_MESSAGE}`, data: additionalData }).end();
+                res.status(200).json({ message: `Actor is using Model Context Protocol. ${HELP_MESSAGE}`, data: getActorRunData() }).end();
             } catch (error) {
                 log.error(`Error in GET ${Routes.ROOT} ${error}`);
                 res.status(500).json({ message: 'Internal Server Error' }).end();
@@ -47,7 +53,10 @@ export function createExpressApp(host: string,
         .get(async (req: Request, res: Response) => {
             try {
                 log.info(`Received GET message at: ${Routes.SSE}`);
-                await mcpServer.processParamsAndUpdateTools(req.url);
+                const tools = await processParamsGetTools(req.url);
+                if (tools) {
+                    mcpServer.updateTools(tools);
+                }
                 transport = new SSEServerTransport(Routes.MESSAGE, res);
                 await mcpServer.connect(transport);
             } catch (error) {
