+++ src/actor/utils.ts
@@ -3,7 +3,9 @@ import { parse } from 'node:querystring';
 import { Actor } from 'apify';
 
 import { processInput } from './input.js';
-import type { ActorRunData, Input } from './types';
+import type { ActorRunData, Input } from './types.js';
+import { addTool, getActorsAsTools, removeTool } from '../tools/index.js';
+import type { ToolWrap } from '../types.js';
 
 export function parseInputParamsFromUrl(url: string): Input {
     const query = url.split('?')[1] || '';
@@ -11,8 +13,21 @@ export function parseInputParamsFromUrl(url: string): Input {
     return processInput(params);
 }
 
-export function isActorStandby(): boolean {
-    return Actor.getEnv().metaOrigin === 'STANDBY';
+/**
+ * Process input parameters and get tools
+ * If URL contains query parameter `actors`, return tools from Actors otherwise return null.
+ * @param url
+ */
+export async function processParamsGetTools(url: string) {
+    const input = parseInputParamsFromUrl(url);
+    let tools: ToolWrap[] = [];
+    if (input.actors) {
+        tools = await getActorsAsTools(input.actors as string[]);
+    }
+    if (input.enableActorAutoLoading) {
+        tools.push(addTool, removeTool);
+    }
+    return tools;
 }
 
 export function getActorRunData(): ActorRunData | null {
