+++ src/index.ts
@@ -1,10 +1,6 @@
 /*
  This file provides essential functions and tools for MCP servers, serving as a library.
+ It should be the only file that is exported from the package.
 */
 
-export { createExpressApp } from './server.js';
 export { ApifyMcpServer } from './mcp-server.js';
-export { getActorAutoLoadingTools, getActorDiscoveryTools } from './tools/index.js';
-export { addActorToTools, discoverActorsTool, getActorsDetailsTool, removeActorFromTools } from './tools/index.js';
-export { getActorsAsTools } from './actors/tools.js';
-export { searchActorsByKeywords } from './actors/search.js';
