--- src/logger.ts
@@ -1,5 +1,0 @@
-import { log } from 'apify';
-
-log.setLevel(log.LEVELS.DEBUG);
-
-export { log };
